Chapter 2 - transfer website into API/pyproject.toml
[tool.poetry]
name = "chapter 2 - transfer website into api"
version = "0.1.0"
description = ""
authors = ["KubaTaba1uga <jakub@buczynski.net.pl>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
